refactor: security files
