+++ src/main/java/org/team24/coursesmanager/config/SecurityConfig.java
@@ -11,7 +11,7 @@ import org.springframework.security.config.http.SessionCreationPolicy;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-import org.team24.coursesmanager.service.PersonDetailsService;
+import org.team24.coursesmanager.security.PersonDetailsService;
 
 
 @EnableWebSecurity
@@ -31,7 +31,7 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
         http.csrf().disable()
                 .authorizeRequests()
                 .antMatchers("/admin").hasRole("TEACHER")
-                .antMatchers("/auth/login", "/auth/registration", "/error").permitAll()
+                .antMatchers("/auth/login", "/auth/registration", "/error", "/**/test/**").permitAll()
                 .anyRequest().authenticated()
                 .and()
                 .sessionManagement()
