+++ src/main/java/org/team24/coursesmanager/controller/TestController.java
@@ -4,18 +4,13 @@ import lombok.AllArgsConstructor;
 import org.modelmapper.ModelMapper;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 import org.team24.coursesmanager.dto.GroupDto;
 import org.team24.coursesmanager.dto.UserReadDto;
 import org.team24.coursesmanager.entity.Group;
-import org.team24.coursesmanager.entity.User;
-import org.team24.coursesmanager.security.PersonDetails;
 import org.team24.coursesmanager.service.GroupService;
 import org.team24.coursesmanager.service.UserService;
 
@@ -48,19 +43,5 @@ public class TestController {
 
     private GroupDto convertToDto(Group group) {
         return modelMapper.map(group, GroupDto.class);
-    }
-
-    @GetMapping("/admin")
-    public String adminPage() {
-        return "admin";
-    }
-
-    @GetMapping("/showUserInfo")
-    @ResponseBody
-    public String showUserInfo() {
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        PersonDetails personDetails = (PersonDetails) authentication.getPrincipal();
-
-        return personDetails.getUsername();
     }
 }
