+++ src/main/java/org/team24/coursesmanager/security/AuthController.java
@@ -1,20 +1,21 @@
-package org.team24.coursesmanager.controller;
+package org.team24.coursesmanager.security;
 
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.validation.BindingResult;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.team24.coursesmanager.dto.AuthenticationDTO;
-import org.team24.coursesmanager.dto.UserDto;
+import org.team24.coursesmanager.dto.UserWriteDto;
 import org.team24.coursesmanager.entity.User;
-import org.team24.coursesmanager.security.JWTUtil;
-import org.team24.coursesmanager.service.RegistrationService;
 import org.team24.coursesmanager.util.UserValidator;
 
 import javax.validation.Valid;
@@ -27,9 +28,7 @@ public class AuthController {
     private final UserValidator userValidator;
     private final JWTUtil jwtUtil;
     private final ModelMapper modelMapper;
-
     private final AuthenticationManager authenticationManager;
-
 
     @Autowired
     public AuthController(RegistrationService registrationService, UserValidator userValidator,
@@ -42,7 +41,7 @@ public class AuthController {
     }
 
     @PostMapping("/registration")
-    public Map<String, String> performRegistration(@RequestBody @Valid UserDto UserDto,
+    public Map<String, String> performRegistration(@RequestBody @Valid UserWriteDto UserDto,
                                                    BindingResult bindingResult) {
         User user = convertToUser(UserDto);
 
@@ -70,8 +69,16 @@ public class AuthController {
         return Map.of("jwt-token", token);
     }
 
+    @GetMapping("/info")
+    public String showUserInfo() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        PersonDetails personDetails = (PersonDetails) authentication.getPrincipal();
 
-    public User convertToUser(UserDto UserDto) {
+        return personDetails.getUsername();
+    }
+
+
+    public User convertToUser(UserWriteDto UserDto) {
         return this.modelMapper.map(UserDto, User.class);
     }
