+++ src/main/java/org/team24/coursesmanager/security/PersonDetailsService.java
@@ -1,4 +1,4 @@
-package org.team24.coursesmanager.service;
+package org.team24.coursesmanager.security;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UserDetails;
@@ -7,7 +7,6 @@ import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 import org.team24.coursesmanager.entity.User;
 import org.team24.coursesmanager.repository.UserRepository;
-import org.team24.coursesmanager.security.PersonDetails;
 
 import java.util.Optional;
